refactor(service): 重构服务层方法，优化地址处理逻辑
- 新增 getAddress 方法，用于将 Address 实体转换为 AddressVo 对象
- 优化地址信息的获取逻辑，提高代码复用性- 修改部分方法返回类型，从 Integer 改为 Long
- 删除未使用的 API 接口实现- 重构部分代码结构，提高可读性和可维护性

+++ src/main/java/com/jiang/mall/service/ICategoryService.java
@@ -1,7 +1,6 @@
 package com.jiang.mall.service;
 
 import com.baomidou.mybatisplus.extension.service.IService;
-import com.jiang.mall.domain.ResponseResult;
 import com.jiang.mall.domain.entity.Category;
 import com.jiang.mall.domain.vo.CategoryVo;
 
@@ -23,7 +22,7 @@ public interface ICategoryService extends IService<Category> {
 
     boolean updateCategory(Category category);
 
-    Integer getCategoryNum();
+    Long getCategoryNum();
 
     boolean deleteCategory(Category category);
 }
